+++ src/lib.rs
@@ -7,6 +7,7 @@
 //! ```
 //! use beckon::beckon;
 //! use serde::{Deserialize, Serialize};
+//! use std::time::Duration;
 //!
 //! #[derive(Serialize, Deserialize)]
 //! pub struct User {
@@ -37,7 +38,7 @@
 //! );
 //!
 //! # async fn example() -> Result<(), Box<dyn std::error::Error>> {
-//! let client = UserApi::new(reqwest::Url::parse("https://api.example.com")?, Some(5000));
+//! let client = UserApi::new(reqwest::Url::parse("https://api.example.com")?, Duration::from_secs(5));
 //! let _users = client.get_users().await?;
 //! let _user = client.get_users_by_id(&UserPath { id: 1 }).await?;
 //! # Ok(())